requirements.txt
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==1.26.3
pillow==10.2.0
pyparsing==3.2.0
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
triton==3.1.0
wheel==0.44.0